requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
datasets==1.18.3
dill==0.3.4
filelock==3.6.0
frozenlist==1.3.0
fsspec==2022.1.0
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.11.1
joblib==1.1.0
multidict==6.0.2
multiprocess==0.70.12.2
numpy==1.21.5
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
protobuf==3.19.4
pyarrow==7.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4
regex==2022.1.18
requests==2.27.1
sacremoses==0.0.47
sentencepiece==0.1.96
six==1.16.0
tokenizers==0.11.5
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
tqdm==4.62.3
transformers @ git+https://github.com/huggingface/transformers@2c2a31ffbcfe03339b1721348781aac4fc05bc5e
typing_extensions==4.1.1
urllib3==1.26.8
wincertstore==0.2
xxhash==2.0.2
yarl==1.7.2
zipp==3.7.0
